DataInterpreter/requirements.txt
duckdb
numpy
pandas
fastapi
uvicorn
langchain
langchain_community
llama-index
llama_index==0.11.22
pydantic
schemas
openpyxl
xlrd
typing-extensions
llama_index.embeddings.huggingface
llama-index-core
ollama
llama-index-llms-langchain
numpy
pdfminer.six
PyMuPDF
pytesseract
Pillow
watchdog